Move Sidebar to sidebar folder
